requirements.txt
dj-database-url==1.0.0
Django==4.1.7
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
drf-yasg==1.21.3
python-decouple==3.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
types-cryptography==3.3.23.1
typing_extensions==4.4.0
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
cloudinary==1.32.0
